Fixed multiple issues (#7)
* Added CreatedDate and ModifiedDate

* Added createddate and modified date, refactored some code

* Multiple Changes

* Fixed MasterFrameOut data update issue

* Fixed header css

* Fixed the billing talbe population

* Limited the text fields to 50

* Fixed table name and made more seed data

+++ Frames/Components/Shared/CreateOrEditWorker.razor.cs
@@ -2,7 +2,7 @@ namespace Frames.Components.Shared;
 
 public partial class CreateOrEditWorker
 {
-    [Parameter] public int WorkerId { get; set; }
+    [Parameter] public long WorkerId { get; set; }
 
     private WorkerDto? _workerDto;
     private readonly List<string> _workerNames = [];
