+++ Frames/Contracts/Repositories/IFrameTypeRepository.cs
@@ -4,7 +4,7 @@ public interface IFrameTypeRepository
 {
     Task<List<FrameType>> GetFrameTypes(bool trackChanges);
     Task<List<string>> GetFrameTypeNames();
-    Task<FrameType?> GetFrameTypeById(int id, bool trackChanges);
+    Task<FrameType?> GetFrameTypeById(long id, bool trackChanges);
     void CreateNewFrameType(FrameType frameType);
     void UpdateFrameType(FrameType frameType);
     void RemoveFrameType(FrameType frameType);
