+++ Frames/Contracts/Services/IFrameTypeService.cs
@@ -4,8 +4,8 @@ public interface IFrameTypeService
 {
     Task<List<FrameTypeDto>> GetFrameTypes();
     Task<List<string>> GetFrameTypeNames();
-    Task<FrameTypeDto> GetFrameTypeById(int frameTypeId);
+    Task<FrameTypeDto> GetFrameTypeById(long frameTypeId);
     Task CreateFrameType(FrameTypeDto newFrameType);
     Task UpdateFrameType(FrameTypeDto frameType);
-    Task RemoveFrameType(int frameTypeId);
+    Task RemoveFrameType(long frameTypeId);
 }
