+++ Frames/Contracts/Services/IMasterFrameOutService.cs
@@ -6,5 +6,5 @@ public interface IMasterFrameOutService
     Task<FrameOutDto> GetFrameOuts(DateOnly date);
     Task CreateOrUpdateFrameOuts(FrameOutDto dto);
     Task RemoveFrameOuts(DateOnly date);
-    Task RemoveFrameOuts(List<int> ids);
+    Task RemoveFrameOuts(List<long> ids);
 }
