+++ Frames/Entities/BillingSummary.cs
@@ -7,7 +7,7 @@ public class BillingSummary : EntityBase
     public decimal Total { get; set; }
     public decimal TotalPaid { get; set; }
     public decimal LastMonth { get; set; }
-    public int BillingId { get; set; }
+    public long BillingId { get; set; }
 
     [ForeignKey(nameof(BillingId))]
     public Billing? Billing { get; set; }
