+++ Frames/Entities/WorkerFrameIn.cs
@@ -6,6 +6,6 @@ public class WorkerFrameIn : EntityBase
 {
     public DateTime Date { get; set; }
     public int NoOfFrames { get; set; }
-    public int WorkerId { get; set; }
+    public long WorkerId { get; set; }
     [ForeignKey(nameof(WorkerId))] public Worker? Worker { get; set; }
 }
