+++ Frames/Entities/WorkerFrameOut.cs
@@ -5,7 +5,7 @@ namespace Frames.Entities;
 public class WorkerFrameOut : EntityBase
 {
     public DateTime Date { get; set; }
-    public int WorkerId { get; set; }
+    public long WorkerId { get; set; }
     [ForeignKey(nameof(WorkerId))] public Worker? Worker { get; set; }
     public List<WorkerFrameOutType> WorkerFrameOutTypes { get; set; } = [];
 }
